+++ ors-engine/src/main/java/org/heigit/ors/routing/graphhopper/extensions/OrsTagParserFactory.java
@@ -1,11 +1,13 @@
 package org.heigit.ors.routing.graphhopper.extensions;
 
 import com.graphhopper.routing.ev.OrsSurface;
+import com.graphhopper.routing.ev.WayType;
 import com.graphhopper.routing.util.parsers.DefaultTagParserFactory;
 import com.graphhopper.routing.util.parsers.TagParser;
 import com.graphhopper.routing.util.parsers.TagParserFactory;
 import com.graphhopper.util.PMap;
 import org.heigit.ors.routing.graphhopper.extensions.util.parsers.OrsSurfaceParser;
+import org.heigit.ors.routing.graphhopper.extensions.util.parsers.WayTypeParser;
 
 public class OrsTagParserFactory implements TagParserFactory {
     DefaultTagParserFactory defaultTagParserFactory = new DefaultTagParserFactory();
@@ -17,6 +19,7 @@ public class OrsTagParserFactory implements TagParserFactory {
         } catch (IllegalArgumentException e) {
             return switch (name) {
                 case OrsSurface.KEY -> new OrsSurfaceParser();
+                case WayType.KEY -> new WayTypeParser();
                 default -> throw e;
             };
         }